requirements.txt
numpy
transformers
torch
streamlit
flask
python-dotenv
google-generativeai